+++ docs/changelog.rst
@@ -11,6 +11,7 @@ CHANGELOG
 * Do not synchronize not-published treks with published but deleted parents
 * Allow to specify portal in touristic content parsers
 * Fix import of type1 in HebergementsSitraParser
+* Fix source and portal missing in shapefile exports
 
 
 2.11.1 (2016-08-17)